% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{economo}
\alias{economo}
\title{Economo & Koskinas Cytoarchitectonic Atlas}
\format{
A \link[ggseg.formats:ggseg_atlas]{ggseg.formats::ggseg_atlas} object (cortical).
}
\usage{
data(economo)
}
\description{
Brain atlas based on the historical Economo & Koskinas 1925
cytoarchitectonic parcellation with 15 regions per hemisphere.
Contains both 2D polygon geometry for \code{\link[ggseg:ggbrain]{ggseg::geom_brain()}} and
3D vertex indices for \code{\link[ggseg3d:ggseg3d]{ggseg3d::ggseg3d()}}.
}
\examples{
data(economo)
economo
}
\references{
C.F. von Economo, G.N. Koskinas; Die Cytoarchitektonik Der
Hirnrinde Des Erwachsenen Menschen; J. Springer (1925)

Pijnenburg et al., NeuroImage, 239, 2021,
\doi{10.1016/j.neuroimage.2021.118274}
}
\concept{ggseg_atlases}
\keyword{datasets}
