% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{economo}
\alias{economo}
\alias{economo_3d}
\title{economo atlas}
\format{
An object of class \code{brain_atlas} of length 4.

An object of class \code{ggseg3d_atlas} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 4 rows and 4 columns.
}
\usage{
economo

economo_3d
}
\description{
Add a short description of this atlas.
}
\examples{
data(economo)
data(economo_3d)
}
\references{
Add reference to the atlas description paper. make sure to add the url to pubmed
    (\href{https://pubmed.ncbi.nlm.nih.gov/xxxxx/}{PubMed})
    
\itemize{
 \item{economo}{ - economo atlas}
 \item{economo_3d}{ - economo 3d mesh atlas}
}
}
\concept{ggseg_atlases ggseg3d_atlases}
\keyword{datasets}
